+++ Frent/World.Create.cs
@@ -28,7 +28,7 @@ partial class World
         //The jit likes to inline the outer create function and not inline
         //the inner functions - benchmarked to improve perf by 10-20%
         var (id, version) = _recycledEntityIds.TryPop(out var v) ? v : new EntityIDOnly(_nextEntityID++, 0);
-        EntityTable[(uint)id] = new(eloc, version);
+        EntityTable[id] = new(eloc, version);
         entity = new Entity(ID, Version, version, id);
         _entityCreated.Invoke(entity);
         return entity;
